# server/controllers/CRUD_db/generated/LineDetailControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE LineDetailControllerGenerated.js PLEASE EDIT ../LineDetailController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import LineDetailModel from "../../../models/CRUD_db/LineDetailModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import LineDetailController from "../LineDetailController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/linedetail`;
    router.post(baseUrl + "", authorize([]), LineDetailController.create);
    router.delete(baseUrl + "/:id", authorize([]), LineDetailController.delete);
    router.get(baseUrl + "/:id", authorize([]), LineDetailController.get);
    router.get(baseUrl + "", authorize([]), LineDetailController.list);
    router.post(baseUrl + "/:id", authorize([]), LineDetailController.update);
  },


  // CRUD METHODS


  /**
  * LineDetailModel.create
  *   @description CRUD ACTION create
  *   @param LineDetail obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await LineDetailModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LineDetailModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id LineDetail
  *
  */
  delete: async (req, res) => {
    try {
      const result = await LineDetailModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LineDetailModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id LineDetail
  *   @returns LineDetail
  *
  */
  get: async (req, res) => {
    try {
      const result = await LineDetailModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LineDetailModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF LineDetail
  *
  */
  list: async (req, res) => {
    try {
      const result = await LineDetailModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * LineDetailModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id LineDetail
  *   @returns LineDetail
  *
  */
  update: async (req, res) => {
    try {
      const result = await LineDetailModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;