# server/controllers/CRUD_db/generated/PurchaseOrderControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE PurchaseOrderControllerGenerated.js PLEASE EDIT ../PurchaseOrderController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import PurchaseOrderModel from "../../../models/CRUD_db/PurchaseOrderModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import PurchaseOrderController from "../PurchaseOrderController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/purchaseorder`;
    router.post(baseUrl + "", authorize([]), PurchaseOrderController.create);
    router.delete(baseUrl + "/:id", authorize([]), PurchaseOrderController.delete);
    router.get(baseUrl + "/findByPoNumber/:key", authorize([]), PurchaseOrderController.findByPoNumber);
    router.get(baseUrl + "/:id", authorize([]), PurchaseOrderController.get);
    router.get(baseUrl + "/:id/getpo_lineDetails", authorize([]), PurchaseOrderController.getpo_lineDetails);
    router.get(baseUrl + "/:id/getpo_webOrderLevel", authorize([]), PurchaseOrderController.getpo_webOrderLevel);
    router.get(baseUrl + "", authorize([]), PurchaseOrderController.list);
    router.post(baseUrl + "/:id", authorize([]), PurchaseOrderController.update);
  },


  // CRUD METHODS


  /**
  * PurchaseOrderModel.create
  *   @description CRUD ACTION create
  *   @param PurchaseOrder obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PurchaseOrderModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id PurchaseOrder
  *
  */
  delete: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PurchaseOrderModel.findByPoNumber
  *   @description CRUD ACTION findByPoNumber
  *   @param Objectid key Id of the resource PoNumber to search
  *
  */
  findByPoNumber: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.findByPoNumber(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PurchaseOrderModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id PurchaseOrder
  *   @returns PurchaseOrder
  *
  */
  get: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PurchaseOrderModel.getpo_lineDetails
  *   @description CRUD ACTION getpo_lineDetails
  *   @param Objectid id ID of PurchaseOrder from po_lineDetails
  *
  */
  getpo_lineDetails: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.getpo_lineDetails(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PurchaseOrderModel.getpo_webOrderLevel
  *   @description CRUD ACTION getpo_webOrderLevel
  *   @param Objectid id ID of PurchaseOrder from po_webOrderLevel
  *
  */
  getpo_webOrderLevel: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.getpo_webOrderLevel(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PurchaseOrderModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF PurchaseOrder
  *
  */
  list: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * PurchaseOrderModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id PurchaseOrder
  *   @returns PurchaseOrder
  *
  */
  update: async (req, res) => {
    try {
      const result = await PurchaseOrderModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;